CodeSamples/Chapter2/requirements.txt
aiofiles==0.6.0
blinker==1.4
click==7.1.2
h11==0.11.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.1
hyperframe==6.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
priority==1.3.0
PyYAML==5.3.1
Quart==0.14.0
toml==0.10.2
Werkzeug==1.0.1
wsproto==1.0.0
